refactor(core): Move files to presentation layer
Moves UI-related files to the presentation layer
for better organization.

test(core): Update pubspec.lock and widget tests

Updates pubspec.lock with dependency changes. Also
updates widget tests to reflect code moves.
