+++ lib/presentation/screens/cart_screen.dart
@@ -4,11 +4,11 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme.dart';
 import 'package:sellio_mobile/core/design_system/constants/assets.dart';
 import 'package:sellio_mobile/core/design_system/themes/sellio_theme_provider.dart';
-import 'package:sellio_mobile/ui/screens/home/home_screen.dart';
 import '../../core/design_system/constants/app_strings.dart';
 import '../../core/design_system/widgets/buttons/button.dart';
 import '../../core/design_system/widgets/cards/productHorizontalCard.dart';
 import '../../core/design_system/widgets/textField.dart';
+import 'home/home_screen.dart';
 
 class CartScreen extends StatefulWidget {
   const CartScreen({super.key});
