+++ lib/presentation/screens/customize_your_product_screen/widgets/GridProduct.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:sellio_mobile/ui/screens/customize_your_product_screen/widgets/CustomProductCard.dart';
+
+import 'CustomProductCard.dart';
 
 class GridProduct extends StatefulWidget {
   const GridProduct({super.key});
