+++ lib/presentation/screens/home/widgets/products_section.dart
