+++ lib/presentation/screens/home/widgets/special_offer/special_offer_card.dart
