+++ lib/presentation/screens/store_details/about_store/models/ContactInfo.dart
