+++ lib/presentation/screens/store_details/about_store/widgets/rating_section.dart
