+++ lib/presentation/screens/store_details/widgets/featured_items_section.dart
